requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
djangorestframework==3.14.0
filelock==3.10.6
idna==3.4
Pillow==9.5.0
platformdirs==3.2.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.3
tzdata==2023.2
urllib3==2.0.2
virtualenv==20.21.0
